refactor(application): 클래스 위치 변경
변경 내용
- WinningTier를 domain 패키지에서 관리하도록 변경
