+++ src/main/java/lotto/service/LottoService.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 import java.util.List;
 import lotto.domain.Lotto;
 import lotto.domain.WinningLotto;
-import lotto.domain.constant.WinningTier;
+import lotto.domain.WinningTier;
 import lotto.utility.RandomGenerator;
 
 public class LottoService {
